% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFunc.R
\name{pow.curve.AB}
\alias{pow.curve.AB}
\title{Power function for proportion test}
\usage{
pow.curve.AB(baseCR, targetCR, nMax)
}
\arguments{
\item{baseCR}{the base conversion rate}

\item{targetCR}{the target conversion rate}

\item{nMax}{max samples}
}
\value{
None
}
\description{
Plots a power function
}
\examples{
pow.curve.AB(0.1,0.15,10000)

}

